make all dataset quality panel default
